#1535 Jars with given size
* Refactor other modules to support long size instead of int

+++ strongbox-storage/strongbox-storage-layout-providers/strongbox-storage-pypi-layout-provider/src/test/java/org/carlspring/strongbox/testing/artifact/PypiArtifactGeneratorStrategy.java
@@ -18,7 +18,7 @@ public class PypiArtifactGeneratorStrategy
     public Path generateArtifact(PypiArtifactGenerator artifactGenerator,
                                  String id,
                                  String version,
-                                 int size,
+                                 long size,
                                  Map<String, Object> attributesMap)
             throws IOException
     {
